refactor(auth): rename authState to authStore and move side effects to use cases
