+++ src/modules/auth/application/usecases/authUseCases.ts
@@ -1,12 +1,29 @@
-import { createRoot } from 'solid-js'
+import { createEffect, createRoot } from 'solid-js'
 
 import { createAuthService } from '~/modules/auth/application/services/authService'
-import { createAuthStore } from '~/modules/auth/application/store/authState'
+import { createAuthStore } from '~/modules/auth/application/store/authStore'
+import { showPromise } from '~/modules/toast/application/toastManager'
+import { userUseCases } from '~/modules/user/application/usecases/userUseCases'
 import { GUEST_USER_ID } from '~/shared/guest/guestConstants'
 
 const { authStore, authService } = createRoot(() => {
   const authStore = createAuthStore()
   const authService = createAuthService(authStore)
+
+  createEffect(() => {
+    const state = authStore.authState()
+    showPromise(
+      userUseCases
+        .fetchUser(state.user?.id ?? '')
+        .then(userUseCases.forceSwitchToUser_unsafe),
+      {
+        loading: 'Carregando dados do usuário...',
+        error: 'Falha ao carregar dados do usuário',
+      },
+      { context: 'background' },
+    ).catch(() => {})
+  })
+
   return { authStore, authService }
 })
